feat: Renamed Search to Device Discovery
The Search name seamed weird. I think DeviceDiscovery is much better.
